+++ tests/modeling/test_plots.py
@@ -1,4 +1,5 @@
 from pharmpy.modeling import (
+    plot_dv_vs_ipred,
     plot_individual_predictions,
     plot_iofv_vs_iofv,
     plot_transformed_eta_distributions,
@@ -27,4 +28,11 @@ def test_plot_transformed_eta_distributions(load_model_for_test, testdata):
     pe = res.parameter_estimates
     ie = res.individual_estimates
     plot = plot_transformed_eta_distributions(model, pe, ie)
+    assert plot
+
+
+def test_plot_dv_vs_ipred(load_model_for_test, testdata):
+    model = load_model_for_test(testdata / 'nonmem' / 'pheno_real.mod')
+    res = read_modelfit_results(testdata / 'nonmem' / 'pheno_real.mod')
+    plot = plot_dv_vs_ipred(model, res.predictions)
     assert plot